requirements.txt
streamlit
pandas
scikit-learn
librosa
joblib
pyaudio
streamlit-audiorecorder